fix minor type things pointed out by g++
git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@393 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ kern/kern/segment.c
@@ -37,7 +37,7 @@ segment_copy(struct Segment *src, struct Label *newl, struct Segment **dstp)
     if (r < 0)
 	return r;
 
-    for (int i = 0; i < src->sg_ko.ko_npages; i++) {
+    for (uint64_t i = 0; i < src->sg_ko.ko_npages; i++) {
 	void *srcpg, *dstpg;
 	r = kobject_get_page(&src->sg_ko, i, &srcpg, kobj_ro);
 	if (r < 0)
